# frontend/openapi.json
{"openapi": "3.1.0", "info": {"title": "housegen", "version": "0.1.0"}, "paths": {"/api/v1/projects": {"get": {"tags": ["projects"], "summary": "List Projects", "operationId": "list_projects", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ProjectSummaryOut"}, "type": "array", "title": "Response List Projects"}}}}}}, "post": {"tags": ["projects"], "summary": "Create Project", "operationId": "create_project", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_create_project"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/projects/{project_id}": {"get": {"tags": ["projects"], "summary": "Get Project", "operationId": "get_project", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["projects"], "summary": "Delete Project", "operationId": "delete_project", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/projects/{project_id}/generate": {"post": {"tags": ["projects"], "summary": "Generate", "operationId": "generate", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}], "responses": {"202": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JobOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/projects/{project_id}/modify": {"post": {"tags": ["projects"], "summary": "Modify", "operationId": "modify", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ModifyRequest"}}}}, "responses": {"202": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JobOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/projects/{project_id}/jobs": {"get": {"tags": ["projects"], "summary": "List Jobs", "operationId": "list_jobs", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/JobOut"}, "title": "Response List Jobs"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/projects/{project_id}/jobs/{job_id}": {"get": {"tags": ["projects"], "summary": "Get Job", "operationId": "get_job", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}, {"name": "job_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Job Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JobOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/projects/{project_id}/jobs/{job_id}/events": {"get": {"tags": ["projects"], "summary": "Job Events", "operationId": "job_events", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}, {"name": "job_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Job Id"}}, {"name": "after", "in": "query", "required": false, "schema": {"type": "integer", "default": 0, "title": "After"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/JobEventOut"}, "title": "Response Job Events"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/projects/{project_id}/chat": {"get": {"tags": ["projects"], "summary": "Chat History", "operationId": "chat_history", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/ChatMessageOut"}, "title": "Response Chat History"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/projects/{project_id}/versions/{number}/restore": {"post": {"tags": ["projects"], "summary": "Restore Version", "operationId": "restore_version", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}, {"name": "number", "in": "path", "required": true, "schema": {"type": "integer", "title": "Number"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/projects/{project_id}/files": {"get": {"tags": ["projects"], "summary": "Scene Files", "operationId": "scene_files", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}, {"name": "version", "in": "query", "required": false, "schema": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Version"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SceneFilesOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/health": {"get": {"tags": ["meta"], "summary": "Health", "operationId": "health", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": {"type": "string"}, "type": "object", "title": "Response Health"}}}}}}}}, "components": {"schemas": {"Body_create_project": {"properties": {"name": {"type": "string", "maxLength": 200, "minLength": 1, "title": "Name"}, "plan": {"type": "string", "contentMediaType": "application/octet-stream", "title": "Plan", "description": "PDF plan set"}, "photos": {"items": {"type": "string", "contentMediaType": "application/octet-stream"}, "type": "array", "title": "Photos", "description": "fa\u00e7ade photos"}, "sides": {"items": {"type": "string"}, "type": "array", "title": "Sides", "description": "side per photo, same order: north|south|east|west|other"}}, "type": "object", "required": ["name", "plan", "photos", "sides"], "title": "Body_create_project"}, "ChatMessageOut": {"properties": {"id": {"type": "string", "title": "Id"}, "role": {"type": "string", "title": "Role"}, "content": {"type": "string", "title": "Content"}, "job_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Job Id"}, "version_number": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Version Number"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["id", "role", "content", "job_id", "version_number", "created_at"], "title": "ChatMessageOut"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "JobEventOut": {"properties": {"seq": {"type": "integer", "title": "Seq"}, "type": {"type": "string", "title": "Type"}, "payload": {"additionalProperties": true, "type": "object", "title": "Payload"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["seq", "type", "payload", "created_at"], "title": "JobEventOut"}, "JobOut": {"properties": {"id": {"type": "string", "title": "Id"}, "project_id": {"type": "string", "title": "Project Id"}, "kind": {"type": "string", "title": "Kind"}, "status": {"type": "string", "title": "Status"}, "request_text": {"type": "string", "title": "Request Text"}, "error": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Error"}, "result_version": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Result Version"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "finished_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Finished At"}}, "type": "object", "required": ["id", "project_id", "kind", "status", "request_text", "error", "result_version", "created_at", "finished_at"], "title": "JobOut"}, "ModifyRequest": {"properties": {"message": {"type": "string", "maxLength": 4000, "minLength": 1, "title": "Message"}}, "type": "object", "required": ["message"], "title": "ModifyRequest"}, "PhotoOut": {"properties": {"id": {"type": "string", "title": "Id"}, "side": {"type": "string", "enum": ["north", "south", "east", "west", "other"], "title": "Side"}, "original_name": {"type": "string", "title": "Original Name"}, "url": {"type": "string", "title": "Url"}}, "type": "object", "required": ["id", "side", "original_name", "url"], "title": "PhotoOut"}, "ProjectOut": {"properties": {"id": {"type": "string", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "status": {"type": "string", "title": "Status"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "plan_pages": {"type": "integer", "title": "Plan Pages"}, "current_version": {"type": "integer", "title": "Current Version"}, "photos": {"items": {"$ref": "#/components/schemas/PhotoOut"}, "type": "array", "title": "Photos"}, "versions": {"items": {"$ref": "#/components/schemas/SceneVersionOut"}, "type": "array", "title": "Versions"}, "scene_url": {"type": "string", "title": "Scene Url"}, "plan_page_urls": {"items": {"type": "string"}, "type": "array", "title": "Plan Page Urls"}}, "type": "object", "required": ["id", "name", "status", "created_at", "plan_pages", "current_version", "photos", "versions", "scene_url", "plan_page_urls"], "title": "ProjectOut"}, "ProjectSummaryOut": {"properties": {"id": {"type": "string", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "status": {"type": "string", "title": "Status"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "current_version": {"type": "integer", "title": "Current Version"}, "thumbnail_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Thumbnail Url"}}, "type": "object", "required": ["id", "name", "status", "created_at", "current_version", "thumbnail_url"], "title": "ProjectSummaryOut"}, "SceneFileOut": {"properties": {"path": {"type": "string", "title": "Path"}, "content": {"type": "string", "title": "Content"}}, "type": "object", "required": ["path", "content"], "title": "SceneFileOut"}, "SceneFilesOut": {"properties": {"version": {"type": "integer", "title": "Version"}, "files": {"items": {"$ref": "#/components/schemas/SceneFileOut"}, "type": "array", "title": "Files"}}, "type": "object", "required": ["version", "files"], "title": "SceneFilesOut"}, "SceneVersionOut": {"properties": {"id": {"type": "string", "title": "Id"}, "number": {"type": "integer", "title": "Number"}, "kind": {"type": "string", "title": "Kind"}, "label": {"type": "string", "title": "Label"}, "summary": {"type": "string", "title": "Summary"}, "critic_score": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Critic Score"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "scene_url": {"type": "string", "title": "Scene Url"}, "render_urls": {"additionalProperties": {"type": "string"}, "type": "object", "title": "Render Urls"}}, "type": "object", "required": ["id", "number", "kind", "label", "summary", "critic_score", "created_at", "scene_url", "render_urls"], "title": "SceneVersionOut"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}